ziyara-reco/requirements.txt
flask
scikit-learn
SQLAlchemy
PyMySQL
python-dotenv
